Moving a few bam tools to a new location for umi tools.
