+++ src/test/resources/com/fulcrumgenomics/umi/annotate_umis.sam
